hw3/appengine/requirements.txt
fastapi
uvicorn[standard]
google-cloud-aiplatform
google-cloud-pubsub
google-cloud-firestore
passlib[bcrypt]
sqlalchemy
pymysql
cryptography
pyjwt
pydantic[email]
pydantic-settings